npm run prettify

+++ contracts/IERC721.sol
@@ -1,25 +1,37 @@
 pragma solidity ^0.5.0;
 
-import "./IERC165.sol";
+import './IERC165.sol';
 
 /**
  * @dev Required interface of an ERC721 compliant contract.
  */
 contract IERC721 is IERC165 {
-    event Transfer(address indexed from, address indexed to, uint256 indexed tokenId);
-    event Approval(address indexed owner, address indexed approved, uint256 indexed tokenId);
-    event Revoke(address indexed owner, address indexed revoked, uint256 indexed tokenId);
-    /**
+  event Transfer(
+    address indexed from,
+    address indexed to,
+    uint256 indexed tokenId
+  );
+  event Approval(
+    address indexed owner,
+    address indexed approved,
+    uint256 indexed tokenId
+  );
+  event Revoke(
+    address indexed owner,
+    address indexed revoked,
+    uint256 indexed tokenId
+  );
+  /**
      * @dev Returns the number of NFTs in `owner`'s account.
      */
-    function balanceOf(address owner) public view returns (uint256 balance);
+  function balanceOf(address owner) public view returns (uint256 balance);
 
-    /**
+  /**
      * @dev Returns the owner of the NFT specified by `tokenId`.
      */
-    function ownerOf(uint256 tokenId) public view returns (address owner);
+  function ownerOf(uint256 tokenId) public view returns (address owner);
 
-    /**
+  /**
      * @dev Transfers a specific NFT (`tokenId`) from one account (`from`) to
      * another (`to`).
      *
@@ -27,12 +39,18 @@ contract IERC721 is IERC165 {
      * - If the caller is not `from`, it must be approved to move this NFT by
      * either {approve} or {setApprovalForAll}.
      */
-    function transferFrom(address from, address to, uint256 tokenId) public;
-    function approve(address to, uint256 tokenId) public;
-    //function lastApproved(uint256 tokenId) public view returns (address operator);
+  function transferFrom(address from, address to, uint256 tokenId) public;
+  function approve(address to, uint256 tokenId) public;
+  //function lastApproved(uint256 tokenId) public view returns (address operator);
 
-    function isApprovedByOwner(address owner, address operator) public view returns (bool);
-    function isApprovedForNFT(address operator, uint idMF) public view returns (bool);
-    function revoke(address to, uint256 tokenId) public;
+  function isApprovedByOwner(address owner, address operator)
+    public
+    view
+    returns (bool);
+  function isApprovedForNFT(address operator, uint256 idMF)
+    public
+    view
+    returns (bool);
+  function revoke(address to, uint256 tokenId) public;
 
 }
